fix: extract hero channel ticker component
